+++ packages/shared/src/tools/index.ts
@@ -3,4 +3,3 @@ export * from './getEnv'
 export * from './ILogger'
 export * from './Logger'
 export * from './LogThrottler'
-export * from './types'
